Fix some glitches
In particular, a crash which happened due a village generating in an
intersection chunk, and a glitch which caused snow layers to not
generate correctly if they are spawned above y62. I'm still not sure
what the cause of that was, but I will try to do more research and find
a definitive solution (very minor changes to this code seems to vastly
affect the outcome, in non forseen ways).