Start using codecs for config serialization
This also removes the mixin for preventing placed features from
appearing in overridden areas in chunks. This may be re-added later, but
should probably use a different method with more context that could be
used to garner what dimension the placed features are being used in, as
each dimension will now have its own unique equation.