Add support for renaming variables without errors
Before this, if one inputted a variable definition like x = x + 1, a
function would be defined like x(x) = x + 1. mXparser encounters an
error when evaluating the function because the name "x" is used twice,
so instead now function names are ensured to not collide with input
variables, so new function will be made as xa(x) instead.